Observer Pattern Implemented
In using the observer design pattern I intend to make communication between classes more streamlined and easier to maintain. I figured since the second lab is an event planner and involves a lot of changes to be made in one class that affects another that this pattern would fit it well. I had to add an “EventObserver” interface and “EventLogger” class, along with methods in the “Event” class to implement them. I also made changes to the “EventTestser” class to show the results.