Command Pattern Implemented
I also implemented the command pattern along with the observer pattern. I was hoping to achieve observers that are notified each time a command is performed in altering the details of an event. To do this I needed to add a “Command” interface and “SetDateTime”, “SetNameCommand”, and “CommandInvoker” classes. As well as further modifying the “EventTester” class to implement my changes.